in progress

+++ Source/SmartHub/SmartHub.UWP.Plugins.Wemos/Core/WemosLineValueType.cs
@@ -4,7 +4,7 @@
     {
         String,
         Boolean,
-        Decimal,
+        Integer,
         Float,
         Custom
     }
